more styling and renaming
