+++ src/pages/registration/contributors_tab/components/AuthorsList.tsx
